refactor(AnalyticsDialog): consolidate files
