+++ client/src/dialogs/Analytics/AnalyticsDialog.scss
