+++ client/src/dialogs/Analytics/AnalyticsDialog.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { vi } from 'vitest'
 
 import { render } from '~/test/helpers/render'
-import AnalyticsDialog from '../AnalyticsDialog'
+import AnalyticsDialog from './AnalyticsDialog'
 
 // Provide mock capacity data to prevent changes in production data from
 // breaking the expected values of this test
