+++ client/src/dialogs/Analytics/CapacitySources.tsx
@@ -1,10 +1,11 @@
 import React, { useState } from 'react'
 import { FormattedMessage } from 'react-intl'
-import { useSelector, useDispatch } from '../../store/hooks'
-import { DEFAULT_CAPACITY_SOURCE } from '../../streets/constants'
-import { isOwnedByCurrentUser } from '../../streets/owner'
-import { getAllCapacityDataSources } from '../../segments/capacity'
-import { setCapacitySource } from '../../store/actions/street'
+
+import { useSelector, useDispatch } from '~/src/store/hooks'
+import { setCapacitySource } from '~/src/store/actions/street'
+import { DEFAULT_CAPACITY_SOURCE } from '~/src/streets/constants'
+import { isOwnedByCurrentUser } from '~/src/streets/owner'
+import { getAllCapacityDataSources } from '~/src/segments/capacity'
 
 function CapacitySources (): React.ReactElement {
   const source = useSelector(
