+++ client/src/dialogs/Analytics/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AnalyticsDialog'
